Setup separate connection and model file and refactor seed to use
