+++ db/db.js
@@ -1,0 +1,4 @@
+const mongoose = require('mongoose');
+mongoose.connect('mongodb://localhost/front_end_capstone');
+
+module.exports = mongoose.connection;
